refactor(OMN-14291): retire models→nodes import-layering back-edge in omnibase_core
Burn down the frozen `models -> nodes` back-edge in the `.importlinter` oracle
(`core-models-no-upward` contract, OMN-14216). ModelServiceCompute/Reducer/
Effect/Orchestrator under models/services/ each subclass the corresponding
Node* base class (NodeCompute/NodeReducer/NodeEffect/NodeOrchestrator) plus
production mixins — they are not domain models and were never re-exported by
models/services/__init__.py.

Real dependency inversion (relocate to the correct layer), matching the four
prior documented reorg phases already recorded in models/services/__init__.py
(docker, graph, event_bus, orchestrator/workflow models moved out when they
didn't belong):

- Move the four files to nodes/, renamed to the node_* naming convention,
  sibling to the base classes they compose:
    models/services/model_service_compute.py      -> nodes/node_service_compute.py
    models/services/model_service_reducer.py       -> nodes/node_service_reducer.py
    models/services/model_service_effect.py         -> nodes/node_service_effect.py
    models/services/model_service_orchestrator.py   -> nodes/node_service_orchestrator.py
- Update the sole src consumer (infrastructure/infra_bases.py facade) and the
  ~14 test files under tests/unit/models/nodes/node_services/ to import from
  the new nodes.* paths.
- Update docstring usage examples in the moved files, plus two docs files
  (docs/reference/api/nodes.md, docs/guides/node-building/02_NODE_TYPES.md)
  that documented the old import path.
- Record this as Phase 5 in models/services/__init__.py's reorg history.
- Drop the now-unmatched models -> nodes ignore line from .importlinter
  (unmatched_ignore_imports_alerting=ERROR ratchet).

Class names, public API, and behavior are unchanged — pure module-location
move. Per CLAUDE.md, internal module paths are explicitly unstable/no-notice.

Refs OMN-3210
