+++ .importlinter
@@ -76,8 +76,6 @@ ignore_imports =
     # OMN-3210 burn-down: models -> validation (~18 edges)
     omnibase_core.models.** -> omnibase_core.validation
     omnibase_core.models.** -> omnibase_core.validation.**
-    # OMN-3210 burn-down: models -> nodes (~4 edges)
-    omnibase_core.models.** -> omnibase_core.nodes.**
 
 [importlinter:contract:core-midtier-no-top]
 name = Services and validation must not import nodes (direct)
