+++ docs/guides/node-building/02_NODE_TYPES.md
@@ -1141,8 +1141,8 @@ class MyWorkflowOrchestrator(NodeOrchestrator):
         )
 
 # SERVICE WRAPPERS: For production with standard mixins
-from omnibase_core.models.services.model_service_effect import ModelServiceEffect
-from omnibase_core.models.services.model_service_compute import ModelServiceCompute
-from omnibase_core.models.services.model_service_reducer import ModelServiceReducer
-from omnibase_core.models.services.model_service_orchestrator import ModelServiceOrchestrator
+from omnibase_core.nodes.node_service_effect import ModelServiceEffect
+from omnibase_core.nodes.node_service_compute import ModelServiceCompute
+from omnibase_core.nodes.node_service_reducer import ModelServiceReducer
+from omnibase_core.nodes.node_service_orchestrator import ModelServiceOrchestrator
 ```
