+++ src/omnibase_core/nodes/node_service_orchestrator.py
@@ -14,9 +14,15 @@ Pre-composed with essential mixins for production use:
 This service wrapper eliminates boilerplate by pre-wiring commonly used mixins
 for orchestrator nodes that coordinate multi-node workflows and manage dependencies.
 
+Relocated from ``omnibase_core.models.services.model_service_orchestrator``
+(OMN-3210 layering burn-down / OMN-14291): this class subclasses
+``NodeOrchestrator`` and does not belong under ``models/`` (domain data
+models). It lives under ``nodes/`` alongside the other ``Node*`` base classes
+it composes.
+
 Usage Example:
     ```python
-    from omnibase_core.models.services.model_service_orchestrator import ModelServiceOrchestrator
+    from omnibase_core.nodes.node_service_orchestrator import ModelServiceOrchestrator
     from omnibase_core.models.container.model_onex_container import ModelONEXContainer
     from omnibase_core.models.contracts.model_contract_orchestrator import ModelContractOrchestrator
