+++ tests/unit/models/nodes/node_services/test_model_service_compute_integration.py
@@ -29,7 +29,7 @@ from omnibase_core.models.discovery.model_tool_invocation_event import (
     ModelToolInvocationEvent,
 )
 from omnibase_core.models.discovery.model_toolparameters import ModelToolParameters
-from omnibase_core.models.services.model_service_compute import ModelServiceCompute
+from omnibase_core.nodes.node_service_compute import ModelServiceCompute
 
 
 class ComputeNodeForIntegrationTest(ModelServiceCompute):
