+++ tests/unit/models/nodes/node_services/test_model_service_effect_invocation.py
@@ -26,7 +26,7 @@ from omnibase_core.models.discovery.model_tool_response_event import (
 )
 from omnibase_core.models.discovery.model_toolparameters import ModelToolParameters
 from omnibase_core.models.errors.model_onex_error import ModelOnexError
-from omnibase_core.models.services.model_service_effect import ModelServiceEffect
+from omnibase_core.nodes.node_service_effect import ModelServiceEffect
 
 
 @pytest.mark.timeout(60)
