+++ tests/unit/models/nodes/node_services/test_model_service_orchestrator_invocation.py
@@ -24,7 +24,7 @@ from omnibase_core.models.discovery.model_tool_response_event import (
     ModelToolResponseEvent,
 )
 from omnibase_core.models.discovery.model_toolparameters import ModelToolParameters
-from omnibase_core.models.services.model_service_orchestrator import (
+from omnibase_core.nodes.node_service_orchestrator import (
     ModelServiceOrchestrator,
 )
