+++ tests/unit/models/nodes/node_services/test_model_service_reducer_health.py
@@ -37,7 +37,7 @@ from omnibase_core.models.discovery.model_tool_invocation_event import (
     ModelToolInvocationEvent,
 )
 from omnibase_core.models.discovery.model_toolparameters import ModelToolParameters
-from omnibase_core.models.services.model_service_reducer import ModelServiceReducer
+from omnibase_core.nodes.node_service_reducer import ModelServiceReducer
 
 # ============================================================================
 # Test Node Implementation
