+++ tests/unit/models/nodes/node_services/test_model_service_reducer_integration.py
@@ -34,7 +34,7 @@ from omnibase_core.models.contracts.model_contract_reducer import ModelContractR
 from omnibase_core.models.discovery.model_tool_invocation_event import (
     ModelToolInvocationEvent,
 )
-from omnibase_core.models.services.model_service_reducer import ModelServiceReducer
+from omnibase_core.nodes.node_service_reducer import ModelServiceReducer
 
 
 class ReducerNodeForIntegrationTest(ModelServiceReducer):
