+++ tests/unit/models/nodes/node_services/test_model_service_reducer_lifecycle.py
@@ -20,7 +20,7 @@ from omnibase_core.models.discovery.model_tool_invocation_event import (
     ModelToolInvocationEvent,
 )
 from omnibase_core.models.errors.model_onex_error import ModelOnexError
-from omnibase_core.models.services.model_service_reducer import ModelServiceReducer
+from omnibase_core.nodes.node_service_reducer import ModelServiceReducer
 
 
 @pytest.mark.unit
